requirements.txt
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.8.2
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
daphne==4.0.0
distlib==0.3.8
Django==5.2.6
django-crispy-forms==2.3
filelock==3.13.1
Flask==3.0.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
importlib_resources==6.5.2
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.1.2
packaging==24.1
pillow==11.3.0
platformdirs==4.2.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
pyserial==3.5
python-barcode==0.16.1
python-dotenv==1.1.1
python-escpos==3.1
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.4.4
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2024.1
virtualenv==20.25.0
Werkzeug==3.0.3
whitenoise==6.7.0
zope.interface==8.0.1
